Change location of HttpResultMapper
